requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
psycopg2-binary==2.9.7
spacy==3.6.1
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
celery==5.3.1
redis==4.6.0
gunicorn==21.2.0
Jinja2==3.1.2
